vibrational_spin_op_builder.py completed. vibrational_label_builder.py introduced.
