+++ qiskit_nature/problems/second_quantization/vibrational/vibrational_label_builder.py
@@ -9,38 +9,8 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
-""" Spin operator builder. """
-
-from qiskit_nature import WatsonHamiltonian
-from qiskit_nature.components.bosonic_bases import HarmonicBasis
-
-
-# TODO complete once a relevant Vibrational SpinOp class created
-def build_spin_op(watson_hamiltonian: WatsonHamiltonian, basis_size, truncation_order):
-    """
-    Builds a spin operator based on a WatsonHamiltonian object.
-
-    Args:
-        watson_hamiltonian (WatsonHamiltonian): WatsonHamiltonian instance.
-        basis_size: size of a basis
-        truncation_order: order at which an n-body expansion is truncated
-
-    Returns:
-        SpinOp: SpinOp built from a WatsonHamiltonian object.
-    """
-
-    num_modes = watson_hamiltonian.num_modes
-    basis_size = [basis_size] * num_modes
-    boson_hamilt_harm_basis = HarmonicBasis(watson_hamiltonian,  # type: ignore
-                                            basis_size, truncation_order).convert()
-    all_labels = _create_labels(boson_hamilt_harm_basis, truncation_order)
-
-    # bos_op = BosonicOp(boson_hamilt_harm_basis) #TODO switch to Vibrational Spin Op
-
-    return False
-
-
-def _create_labels(boson_hamilt_harm_basis, truncation_order):
+def create_labels(boson_hamilt_harm_basis, truncation_order):
+    # TODO docs
     all_labels = []
     for num_body in range(truncation_order):
         num_body_data = boson_hamilt_harm_basis[num_body]
