+++ qiskit_nature/problems/second_quantization/vibrational/vibrational_problem.py
@@ -16,7 +16,7 @@ from qiskit_nature import WatsonHamiltonian
 from qiskit_nature.drivers import BosonicDriver
 from qiskit_nature.operators.second_quantization import SecondQuantizedOp
 from qiskit_nature.problems.second_quantization.base_problem import BaseProblem
-from qiskit_nature.problems.second_quantization.vibrational.spin_op_builder import build_spin_op
+from qiskit_nature.problems.second_quantization.vibrational.vibrational_spin_op_builder import build_vibrational_spin_op
 from qiskit_nature.transformers import BaseTransformer
 
 
@@ -46,8 +46,8 @@ class VibrationalProblem(BaseProblem):
         """
         watson_hamiltonian = self.driver.run()
         watson_hamiltonian_transformed = self._transform(watson_hamiltonian)
-        bosonic_op = build_spin_op(watson_hamiltonian_transformed, self.basis_size,
-                                   self.truncation_order)
+        bosonic_op = build_vibrational_spin_op(watson_hamiltonian_transformed, self.basis_size,
+                                               self.truncation_order)
 
         second_quantized_ops_list = [SecondQuantizedOp([bosonic_op])]
