+++ test/problems/second_quantization/vibrational/test_vibrational_label_builder.py
@@ -9,15 +9,21 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
-"""Tests Fermionic Operator builder."""
+from qiskit_nature.components.bosonic_bases import HarmonicBasis
+from qiskit_nature.drivers import GaussianForcesDriver
+from qiskit_nature.problems.second_quantization.vibrational.vibrational_label_builder import \
+    create_labels
+
+"""Tests Vibrational Label Builder."""
 from test import QiskitNatureTestCase
 from qiskit_nature.components.bosonic_bases import HarmonicBasis
-from qiskit_nature.problems.second_quantization.vibrational.spin_op_builder import _create_labels
+from qiskit_nature.problems.second_quantization.vibrational.vibrational_label_builder import \
+    create_labels
 from qiskit_nature.drivers import GaussianForcesDriver
 
 
-class TestVibrationalSpinOpBuilder(QiskitNatureTestCase):
-    """Tests Vibrational Spin Op builder."""
+class TestVibrationalLabelBuilder(QiskitNatureTestCase):
+    """Tests Vibrational Label Builder."""
 
     def test_create_labels(self):
         """Tests that correct labels are built."""
@@ -158,5 +164,5 @@ class TestVibrationalSpinOpBuilder(QiskitNatureTestCase):
         basis_size = [basis_size] * num_modes
         boson_hamilt_harm_basis = HarmonicBasis(watson_hamiltonian,  # type: ignore
                                                 basis_size, truncation_order).convert()
-        labels = _create_labels(boson_hamilt_harm_basis, truncation_order)
+        labels = create_labels(boson_hamilt_harm_basis, truncation_order)
         assert labels == expected_labels
